requirements.txt
accelerate==0.32.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.8.0.post1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.5
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.8
langchain-community==0.2.7
langchain-core==0.2.19
langchain-google-genai==1.0.7
langchain-openai==0.1.16
langchain-text-splitters==0.2.2
langsmith==0.1.86
MarkupSafe==2.1.5
marshmallow==3.21.3
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
openai==1.35.14
orjson==3.10.6
packaging==24.1
pip==24.1.2
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rsa==4.9
safetensors==0.4.3
sniffio==1.3.1
SQLAlchemy==2.0.31
sympy==1.13.0
tbb==2021.13.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
typing-inspect==0.9.0
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
